fix：修复图片剪切在某些尺寸下由于没有使用浮点数计算导致尺寸计算错误的bug

+++ jpress/jpress-commons/src/main/java/io/jpress/utils/ImageUtils.java
@@ -100,7 +100,7 @@ public class ImageUtils {
 
 		Rectangle rect = null;
 
-		if (w / h > srcWidth / srcHeight) {
+		if ((float) w / h > (float) srcWidth / srcHeight) {
 			h = h * srcWidth / w;
 			w = srcWidth;
 			rect = new Rectangle(0, (srcHeight - h) / 2, w, h);
